+++ __tests__/integration/queries.test.js
@@ -5,6 +5,8 @@ const { readdirSync, readFile: rawReadFile } = require("fs");
 const { resolve: resolvePath } = require("path");
 const { printSchema } = require("graphql/utilities");
 const debug = require("debug")("graphile-build:schema");
+const { PgConnectionArgCondition } = require("graphile-build-pg");
+const CustomOperatorsPlugin = require("./../customOperatorsPlugin");
 
 function readFile(filename, encoding) {
   return new Promise((resolve, reject) => {
@@ -34,30 +36,31 @@ beforeAll(() => {
       simpleCollections,
       nullAndEmptyAllowed,
       additionalInsensitiveOperators,
+      addConnectionFilterOperator,
     ] = await Promise.all([
       createPostGraphileSchema(pgClient, ["p"], {
-        skipPlugins: [require("graphile-build-pg").PgConnectionArgCondition],
+        skipPlugins: [PgConnectionArgCondition],
         appendPlugins: [require("../../index.js")],
       }),
       createPostGraphileSchema(pgClient, ["p"], {
-        skipPlugins: [require("graphile-build-pg").PgConnectionArgCondition],
+        skipPlugins: [PgConnectionArgCondition],
         appendPlugins: [require("../../index.js")],
         dynamicJson: true,
       }),
       createPostGraphileSchema(pgClient, ["p"], {
-        skipPlugins: [require("graphile-build-pg").PgConnectionArgCondition],
+        skipPlugins: [PgConnectionArgCondition],
         appendPlugins: [require("../../index.js")],
         graphileBuildOptions: {
           connectionFilterRelations: true,
         },
       }),
       createPostGraphileSchema(pgClient, ["p"], {
-        skipPlugins: [require("graphile-build-pg").PgConnectionArgCondition],
+        skipPlugins: [PgConnectionArgCondition],
         appendPlugins: [require("../../index.js")],
         simpleCollections: "only",
       }),
       createPostGraphileSchema(pgClient, ["p"], {
-        skipPlugins: [require("graphile-build-pg").PgConnectionArgCondition],
+        skipPlugins: [PgConnectionArgCondition],
         appendPlugins: [require("../../index.js")],
         graphileBuildOptions: {
           connectionFilterAllowNullInput: true,
@@ -65,11 +68,15 @@ beforeAll(() => {
         },
       }),
       createPostGraphileSchema(pgClient, ["p"], {
-        skipPlugins: [require("graphile-build-pg").PgConnectionArgCondition],
+        skipPlugins: [PgConnectionArgCondition],
         appendPlugins: [require("../../index.js")],
         graphileBuildOptions: {
           connectionFilterAdditionalInsensitiveOperators: true,
         },
+      }),
+      createPostGraphileSchema(pgClient, ["p"], {
+        skipPlugins: [PgConnectionArgCondition],
+        appendPlugins: [require("../../index.js"), CustomOperatorsPlugin],
       }),
     ]);
     debug(printSchema(normal));
@@ -80,6 +87,7 @@ beforeAll(() => {
       simpleCollections,
       nullAndEmptyAllowed,
       additionalInsensitiveOperators,
+      addConnectionFilterOperator,
     };
   });
 
@@ -107,6 +115,8 @@ beforeAll(() => {
           // some specific fixtures against a schema configured slightly
           // differently.
           const schemas = {
+            "addConnectionFilterOperator.graphql":
+              gqlSchemas.addConnectionFilterOperator,
             "additionalInsensitiveOperatorsTrue.citext.graphql":
               gqlSchemas.additionalInsensitiveOperators,
             "additionalInsensitiveOperatorsTrue.text.graphql":